improve utils module
