+++ qulab_toolbox/utils/_d2c.py
@@ -1,0 +1,160 @@
+from blinker import Signal
+import numpy as np
+
+class d2c(object):
+    '''dict to class
+    
+    将字典的调用方式转化为类属性的调用方式；
+    字典的关键字须符合类属性命名的规范，可以赋值
+    '''
+    
+    def __init__(self,d):
+        '''传入一个字典或者d2c类的实例，构造一个d2c实例'''
+        if isinstance(d,d2c):
+            d=d.todict()
+        assert isinstance(d,dict)
+        for k,v in d.items():
+            _v=d2c(v) if isinstance(v,dict) else v           
+            self.__dict__.update({k:_v})
+            
+    def todict(self):
+        '''从d2c实例提取出原字典'''
+        d={}
+        for k,v in self.__dict__.items():
+            _v=v.todict() if isinstance(v,d2c) else v
+            d.update({k:_v})
+        return d
+
+    def get(self,keys,splitsymbol='.'):
+        '''根据关键词连接的字符串或者列表，获取对应的值
+        
+        Parameters:
+            keys: 使用符号串联key的字符串，比如'pulse.width.value';
+                或者key顺序列表，比如['pulse','width','value']
+            splitsymbol: 字符串的分割符号，默认为 '.'
+        Return:
+            返回键值串对应的值
+        '''
+        keys=keys.split(splitsymbol) if isinstance(keys,str) else keys
+        value=self
+        for k in keys:
+            value=getattr(value,k)
+        ##
+        # value=getattr(self,keys.pop(0))
+        # if keys:
+        #     value=value.get(keys) #递归
+        return value
+
+    def set(self,keys,value,check=False,splitsymbol='.'):
+        '''根据关键词连接的字符串或者列表，设置对应的值
+        
+        Parameters:
+            keys: 使用符号串联key的字符串，比如'pulse.width.value';
+                或者key顺序列表，比如['pulse','width','value']
+            value: 待设入的值
+            check: bool, 是否检查设入值与原值相等
+            splitsymbol: 字符串的分割符号，默认为 '.'
+        Return:
+            bool, True 表示value改变并且已设置，False 表示未变
+        '''
+        if check:
+            if is_equal(value,self.get(keys)):
+                return False
+        keys=keys.split(splitsymbol) if isinstance(keys,str) else keys
+        ins=self
+        for k in keys[:-1]:
+            ins=getattr(ins,k)
+        assert hasattr(ins,keys[-1]),f'关键词 {keys[-1]} 不存在!'
+        setattr(ins,keys[-1],value)
+        ## 
+        # if len(keys)>1:
+        #     getattr(self,keys.pop(0)).set(keys,value) ##递归
+        # else:
+        #     assert hasattr(self,keys[0]) ##必须是字典中已有的关键词
+        #     setattr(self,keys[0],value)
+        return True
+
+def is_equal(v1,v2):
+    '''深入地比较v1,v2是否相等'''
+    try:
+        assert_equal(v1,v2)
+        return True
+    except AssertionError:
+        return False
+
+def assert_equal(v1,v2):
+    '''断言v1,v2相等'''
+    try:
+        if isinstance(v1,np.ndarray) or isinstance(v1,np.ndarray):
+            assert np.all(v1==v2)
+        else:
+            assert v1==v2
+    except ValueError: # ValueError: The truth value of an array with more than one element is ambiguous. Use a.any() or a.all()
+        if isinstance(v1,dict) and isinstance(v2,dict):
+            assert not set(v1.keys())^set(v2.keys()),Exception('keys not equal !')
+            for k in v1.keys():
+                _v1,_v2=v1[k],v2[k]
+                assert_equal(_v1,_v2)
+        elif isinstance(v1,(list,tuple,set)) and isinstance(v2,(list,tuple,set)):
+            for _v1,_v2 in zip(v1,v2):
+                assert_equal(_v1,_v2)
+        else:
+            assert False
+
+
+class cons_d2c(object):
+    '''包含约束的d2c类'''
+
+    def __init__(self,d,constrains,init=True):
+        '''
+        Parameters:
+            d: 一个字典
+            constrains: 约束字典，例如，
+                {
+                    'k2':(('k1',), lambda v: v*3)
+                    'k3': (('k1','k2'), lambda v1,v2:v1+v2),
+                }
+            init: bool, 是否使用约束条件初始化传入的字典
+        '''
+        self.__d2c_instance=d2c(d)
+        self.sig=Signal()
+        self.constrains=constrains
+        self.subscribers=self.__init_cons(constrains,init=init)
+
+    def __init_cons(self,constrains,init=True):
+        subscribers={}
+        for k,v in constrains.items():
+            deps,func=v
+            subs=_subscrib(k,deps,func,self)
+            for _k in deps:
+                self.sig.connect(subs,sender=_k)
+                if init:
+                    self.sig.send(_k)
+            subscribers.update({k:subs})
+        return subscribers
+        
+    def set(self,k,v):
+        _changed=self.__d2c_instance.set(k,v,check=True)
+        if _changed:
+            self.sig.send(k)
+            
+    def get(self,k):
+        return self.__d2c_instance.get(k)
+    
+    def __getattr__(self,item):
+        return getattr(self.__d2c_instance,item)
+
+def _subscrib(key, deps=(), func=None, cons_d2c_ins=None):
+    '''根据需要的参量产生一个订阅函数
+    Parameters:
+        key,deps,func: 分别为目标的key、依赖的key列表或元组、约束函数，比如
+            ('k3',('k1','k2'),lambda v1,v2:v1+v2)
+        cons_d2c_ins: 上面cons_d2c的一个实例
+    Return:
+        返回一个订阅函数
+    '''
+    def subscriber(send):
+        arg = [cfgmgr.get(_k) for _k in deps]
+        value=func(*arg)
+        cfgmgr.set(key,value)
+    return subscriber
